Controllers and routes ,error usercontroller->update

+++ Laravel-User-Role-Api-and-Vue/routes/api.php
@@ -2,6 +2,9 @@
 
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\AuthController;
+use App\Http\Controllers\RoleController;
+use App\Http\Controllers\UserController;
 
 /*
 |--------------------------------------------------------------------------
@@ -21,30 +24,49 @@ use Illuminate\Support\Facades\Route;
 /* JWt route */
 
 //Route for authentification
-Route::post('/login', 'AuthController@login');
-Route::post('/register', 'AuthController@register');
+Route::post('/login', [AuthController::class,'login']);
+Route::post('/register', [AuthController::class,'register']);
 
 //Route for logout or refresh tokens
 Route::middleware('jwt.auth')->group(function () {
-    Route::post('/logout', 'AuthController@logout');
-    Route::post('/refresh', 'AuthController@refresh');
+    Route::post('/logout',  [AuthController::class,'logout']);
+    Route::post('/refresh', [AuthController::class,'refresh']);
 });
 
 // Routes for CRUD operations on roles
 Route::middleware('jwt.auth')->group(function () {
-    Route::get('role', 'RoleController@index');
-    Route::get('role/{id}', 'RoleController@show');
-    Route::post('role', 'RoleController@store');
-    Route::put('role/{id}', 'RoleController@update');
-    Route::delete('role/{id}', 'RoleController@destroy');
-    Route::delete('/role/{id}', 'RoleController@destroy')->middleware('auth.jwt', 'check.role:3');
+    Route::get('role', [RoleController::class,'index']);
+    Route::get('role/{id}', [RoleController::class,'show']);
+    Route::post('role', [RoleController::class,'store']);
+    Route::put('role/{id}', [RoleController::class,'update']);
+    // Route::delete('role/{id}', [RoleController::class,'destroy']);
+    Route::delete('/role/{id}', [RoleController::class,'destroy'])->middleware('jwt.auth', 'checkrole:3');
 });
 
 // Routes for CRUD operations on users
 Route::middleware('jwt.auth')->group(function () {
-    Route::get('user', 'UserController@index');
-    Route::get('user/{id}', 'UserController@show');
-    Route::post('user', 'UserController@store');
-    Route::put('user/{id}', 'UserController@update');
-    Route::delete('/user/{id}', 'UserController@destroy')->middleware('auth.jwt', 'check.role:3');
+    Route::get('user', [UserController::class,'index']);
+    Route::get('user/{id}', [UserController::class,'show']);
+    Route::post('user', [UserController::class,'store']);
+    Route::put('user/{id}', [UserController::class,'update']);
+    Route::delete('/user/{id}', [UserController::class,'destroy'])->middleware('jwt.auth', 'checkrole:3');
 });
+
+
+
+// //Routes for authentification
+// Route::post('/login', 'AuthController@login');
+// Route::post('/register', 'AuthController@register');
+
+// //Routes for logout or refresh tokens
+// Route::middleware('jwt.auth')->group(function () {
+//     Route::post('/logout', 'AuthController@logout');
+//     Route::post('/refresh', 'AuthController@refresh');
+// });
+
+// //Routes for CRUD operations on roles
+// Route::middleware('jwt.auth')->resource('role', 'RoleController', ['except' => ['create', 'edit']])->middleware('auth.jwt', 'check.role:3');
+
+// //Routes for CRUD operations on users
+// Route::middleware('jwt.auth')->resource('user', 'UserController', ['except' => ['create', 'edit']])->middleware('auth.jwt', 'check.role:3');
+
